requirements.txt
google-api-python-client 
python-dotenv
paho-mqtt